refactor(languages)!: Overhaul language support as dedicated core module type [pt1]
BREAKING CHANGE: Language support has been completely overrhauled;
classes and packages that made direct use of language tooling internals
will need to be refactored. The global `SILE.languageSupport` and module
`core.language` have been completely deprecated. Language modules are
now a first class module type so modifying or extending them should be
much easier. Like document classes, packages, typesetters, shapers,
inputters, outputter, and other module types there is a `languages.base`
prototype. Additionally Unicode specific support is in
`languages.unicode` and most language support modules build on this.
Each language is in a `language.<lang>` module.

Access to the current active language module is provided via the active
typesetter's `typesetter.language` property. Additionally when switching
languages the typesetter will cache all instanciated language support
modules with any associated settings. Access to other languages module
can be achieved by asking the typesetter for it's cache of language
instances. If one hasn't been made for a language it will be
instanceated and put in the cache.
