+++ classes/base.lua
@@ -54,7 +54,6 @@ function class:_init (options)
    if self == options then
       options = {}
    end
-   SILE.languageSupport.loadLanguage("und") -- preload for unlocalized fallbacks
    self:_declareBaseOptions()
    self:declareOptions()
    self:registerRawHandlers()
