+++ core/break.lua
@@ -709,7 +709,7 @@ function lineBreak:doBreak (nodes, hsize, sideways)
          self.threshold = inf_bad
       end
       if self.pass == "second" then
-         self.nodes = SILE.hyphenate(self.nodes)
+         self.nodes = SILE.language:hyphenate(self.nodes)
          SILE.typesetter.state.nodes = self.nodes -- Horrible breaking of separation of concerns here. :-(
       end
       -- 890
