+++ core/deprecations.lua
@@ -214,6 +214,42 @@ SU.utf8codes = function ()
    SU.deprecated("SU.utf8codes", "luautf8.codes", "0.11.0", "0.12.0")
 end
 
+local nolanguageloader = function (_, key)
+   if key == "loadLanguage" then
+      return function (language)
+         SU.deprecated("SILE.languageSupport.<code>", ('require("languages.%s")'):format(language), "0.16.0", "0.17.0")
+         return SILE.languages[language]
+      end
+   end
+end
+
+SILE.languageSupport = {}
+setmetatable(SILE.languageSupport, {
+   __index = nolanguageloader,
+})
+
+local nonodemakers = function (_, key)
+   SU.deprecated(
+      "SILE.nodeMakers.<code>",
+      'require("SILE.language.?")', --TODO
+      "0.15.11",
+      "0.17.0",
+      [[
+      ]]
+   )
+   if key == "loadLanguage" then
+      return function (language)
+         require(("languages.%s"):format(language))
+      end
+   end
+end
+
+SILE.nodeMakers = {}
+setmetatable(SILE.nodeMakers, {
+   __index = nonodemakers,
+})
+
+
 -- luacheck: ignore updatePackage
 -- luacheck: ignore installPackage
 updatePackage = nopackagemanager
