+++ languages/tr/init.lua
@@ -1,13 +1,20 @@
--- Different years of TDK and various publisher style guides differ on this point.
--- Current official guidance suggests dropping the hyphenation mark if the break
--- occurs at an apostrophe (kesme işareti). Some older guidance and some publishers
--- suggest dropping the apostrophe instead.
-SILE.settings:declare({
-   parameter = "languages.tr.replaceApostropheAtHyphenation",
-   type = "boolean",
-   default = false,
-   help = "If enabled, substitute the apostophe for a hyphen at break points, otherwise keep the apostrophe and hide the hyphen.",
-})
+local base = require("languages.base")
+
+local language = pl.class(base)
+language._name = "tr"
+
+function language.declareSettings (_)
+   -- Different years of TDK and various publisher style guides differ on this point.
+   -- Current official guidance suggests dropping the hyphenation mark if the break
+   -- occurs at an apostrophe (kesme işareti). Some older guidance and some publishers
+   -- suggest dropping the apostrophe instead.
+   SILE.settings:declare({
+      parameter = "languages.tr.replaceApostropheAtHyphenation",
+      type = "boolean",
+      default = false,
+      help = "If enabled, substitute the apostophe for a hyphen at break points, otherwise keep the apostrophe and hide the hyphen.",
+   })
+end
 
 -- Quotes may be part of a word in Turkish
 SILE.nodeMakers.tr = pl.class(SILE.nodeMakers.unicode)
