+++ languages/und/init.lua
@@ -1,1 +1,8 @@
-SILE.hyphenator.languages.und = { patterns = {}, exceptions = {} }
+local base = require("languages.base")
+
+local language = pl.class(base)
+language._name = "und"
+
+-- SILE.hyphenator.languages.und = { patterns = {}, exceptions = {} }
+
+return language
